cxx
Module kind

Marker types identifying Rust’s knowledge about an extern C++ type.

These markers are used in the Kind associated type in impls of the ExternType trait. Refer to the documentation of Kind for an overview of their purpose.

Enums

Opaque

An opaque type which cannot be passed or held by value within Rust.

Trivial

A type with trivial move constructor and no destructor, which can therefore be owned and moved around in Rust code without requiring indirection.

Traits

Kind