Module cxx::memory

Less used details of UniquePtr and SharedPtr.

The pointer types themselves are exposed at the crate root.

Re-exports

pub use cxx::SharedPtr;

pub use cxx::UniquePtr;

Traits

SharedPtrTarget

Trait bound for types which may be used as the T inside of a SharedPtr<T> in generic code.

UniquePtrTarget

Trait bound for types which may be used as the T inside of a UniquePtr<T> in generic code.

WeakPtrTarget

Trait bound for types which may be used as the T inside of a WeakPtr<T> in generic code.