Struct cxx::UniquePtr

#[repr(C)]
pub struct UniquePtr<T>
where
    T: UniquePtrTarget,
{ /* private fields */ }

Binding to C++ std::unique_ptr<T, std::default_delete<T>>.

Implementations

impl<T> UniquePtr<T>
where T: UniquePtrTarget,

pub fn null() -> Self

Makes a new UniquePtr wrapping a null pointer.

Matches the behavior of default-constructing a std::unique_ptr.

pub fn new(value: T) -> Self
where T: ExternType<Kind = Trivial>,

Allocates memory on the heap and makes a UniquePtr pointing to it.

pub fn is_null(&self) -> bool

Checks whether the UniquePtr does not own an object.

This is the opposite of std::unique_ptr<T>::operator bool.

pub fn as_ref(&self) -> Option<&T>

Returns a reference to the object owned by this UniquePtr if any, otherwise None.

pub fn as_mut(&mut self) -> Option<Pin<&mut T>>

Returns a mutable pinned reference to the object owned by this UniquePtr if any, otherwise None.

pub fn pin_mut(&mut self) -> Pin<&mut T>

Returns a mutable pinned reference to the object owned by this UniquePtr.

Panics

Panics if the UniquePtr holds a null pointer.

pub fn into_raw(self) -> *mut T

Consumes the UniquePtr, releasing its ownership of the heap-allocated T.

Matches the behavior of std::unique_ptr<T>::release.

pub unsafe fn from_raw(raw: *mut T) -> Self

Constructs a UniquePtr retaking ownership of a pointer previously obtained from into_raw.

Safety

This function is unsafe because improper use may lead to memory problems. For example a double-free may occur if the function is called twice on the same raw pointer.