cxx

Struct CxxString

#[repr(C)]
pub struct CxxString { /* private fields */ }

Binding to C++ std::string.

Invariants

As an invariant of this API and the static analysis of the cxx::bridge macro, in Rust code we can never obtain a CxxString by value. C++’s string requires a move constructor and may hold internal pointers, which is not compatible with Rust’s move behavior. Instead in Rust code we will only ever look at a CxxString through a reference or smart pointer, as in &CxxString or UniquePtr<CxxString>.

Implementations

impl CxxString

pub fn new<T: Private>() -> Self

CxxString is not constructible via new. Instead, use the let_cxx_string! macro.

pub fn len(&self) -> usize

Returns the length of the string in bytes.

Matches the behavior of C++ std::string::size.

pub fn is_empty(&self) -> bool

Returns true if self has a length of zero bytes.

Matches the behavior of C++ std::string::empty.

pub fn as_bytes(&self) -> &[u8]

Returns a byte slice of this string’s contents.

pub fn as_ptr(&self) -> *const u8

Produces a pointer to the first character of the string.

Matches the behavior of C++ std::string::data.

Note that the return type may look like const char * but is not a const char * in the typical C sense, as C++ strings may contain internal null bytes. As such, the returned pointer only makes sense as a string in combination with the length returned by len().

Modifying the string data through this pointer has undefined behavior.

pub fn as_c_str(&self) -> &CStr

Produces a nul-terminated string view of this string’s contents.

Matches the behavior of C++ std::string::c_str.

If this string contains no internal ‘\0’ bytes, then self.as_c_str().count_bytes() == self.len(). But if it does, the CStr only refers to the part of the string up to the first nul byte.

pub fn to_str(&self) -> Result<&str, Utf8Error>

Validates that the C++ string contains UTF-8 data and produces a view of it as a Rust &str, otherwise an error.

pub fn to_string_lossy(&self) -> Cow<'_, str>

If the contents of the C++ string are valid UTF-8, this function returns a view as a Cow::Borrowed &str. Otherwise replaces any invalid UTF-8 sequences with the U+FFFD replacement character and returns a Cow::Owned String.

pub fn clear(self: Pin<&mut Self>)

Removes all characters from the string.

Matches the behavior of C++ std::string::clear.

Note: unlike the guarantee of Rust’s std::string::String::clear, the C++ standard does not require that capacity is unchanged by this operation. In practice existing implementations do not change the capacity but all pointers, references, and iterators into the string contents are nevertheless invalidated.

pub fn reserve(self: Pin<&mut Self>, additional: usize)

Ensures that this string’s capacity is at least additional bytes larger than its length.

The capacity may be increased by more than additional bytes if the implementation chooses, to amortize the cost of frequent reallocations.

The meaning of the argument is not the same as std::string::reserve in C++. The C++ standard library and Rust standard library both have a reserve method on strings, but in C++ code the argument always refers to total capacity, whereas in Rust code it always refers to additional capacity. This API on CxxString follows the Rust convention, the same way that for the length accessor we use the Rust conventional len() naming and not C++ size() or length().

Panics

Panics if the new capacity overflows usize.

pub fn push_str(self: Pin<&mut Self>, s: &str)

Appends a given string slice onto the end of this C++ string.

pub fn push_bytes(self: Pin<&mut Self>, bytes: &[u8])

Appends arbitrary bytes onto the end of this C++ string.